requirements.txt
pytest
flake8
azure-functions